+++ cvat-ui/src/components/file-manager/cloud-storages-tab.tsx
@@ -13,7 +13,7 @@ import { debounce } from 'lodash';
 import Select from 'antd/lib/select';
 import getCore from 'cvat-core-wrapper';
 import { CloudStorage } from 'reducers/interfaces';
-import { AzureProvider, S3Provider } from 'icons';
+import { AzureProvider, GoogleCloudProvider, S3Provider } from 'icons';
 import { ProviderType } from 'utils/enums';
 import CloudStorageFiles from './cloud-storages-files';
 
@@ -125,11 +125,12 @@ export default function CloudStorageTab(props: Props): JSX.Element {
                             <span
                                 className='cvat-cloud-storage-select-provider'
                             >
-                                {_cloudStorage.providerType === ProviderType.AWS_S3_BUCKET ? (
-                                    <S3Provider />
-                                ) : (
-                                    <AzureProvider />
-                                )}
+                                {_cloudStorage.providerType === ProviderType.AWS_S3_BUCKET && <S3Provider />}
+                                {_cloudStorage.providerType === ProviderType.AZURE_CONTAINER && <AzureProvider />}
+                                {
+                                    _cloudStorage.providerType === ProviderType.GOOGLE_CLOUD_STORAGE &&
+                                    <GoogleCloudProvider />
+                                }
                                 {_cloudStorage.displayName}
                             </span>
                         ),