+++ cvat-ui/src/reducers/projects-reducer.ts
@@ -24,6 +24,18 @@ const defaultState: ProjectsState = {
         name: null,
         status: null,
     },
+    tasksGettingQuery: {
+        page: 1,
+        id: null,
+        search: null,
+        owner: null,
+        assignee: null,
+        name: null,
+        status: null,
+        mode: null,
+        projectId: null,
+        ordering: 'subset',
+    },
     activities: {
         deletes: {},
         creates: {
@@ -43,6 +55,10 @@ export default (state: ProjectsState = defaultState, action: AnyAction): Project
                 gettingQuery: {
                     ...defaultState.gettingQuery,
                     ...action.payload.query,
+                },
+                tasksGettingQuery: {
+                    ...defaultState.tasksGettingQuery,
+                    ...action.payload.tasksQuery,
                 },
             };
         case ProjectsActionTypes.GET_PROJECTS: